+++ CHANGELOG.md
@@ -5,7 +5,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.11.0]
 - Andded 'cfnetwork:firewall' anchor
-- Added a new 'cf_get_bind_address' API to properly retrieve iface bind address
+- Added a new 'cfnetwork::bind_address' API to properly retrieve iface bind address
 - Deprecated 'cf_get_iface_address'
 - Added automatic creation of host entries for each interface and additional addresses
 - Added cfnetwork::hosts parameter